# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OrthoVision Treatment Planning Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: patients
    description: Patient management
  - name: cases
    description: Orthodontic case management
  - name: scans
    description: 3D scan management
  - name: dashboard
    description: Dashboard analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Doctor login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseListResponse"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrthoCase"

  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrthoCase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrthoCase"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /cases/{caseId}/scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scans for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"

  /scans/{scanId}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete a scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/case-status-breakdown:
    get:
      operationId: getCaseStatusBreakdown
      tags: [dashboard]
      summary: Get case counts by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [doctor]
      properties:
        doctor:
          $ref: "#/components/schemas/Doctor"

    Doctor:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    Patient:
      type: object
      required: [id, patientCode, fullName, age, gender, mobileNumber, email, createdAt]
      properties:
        id:
          type: integer
        patientCode:
          type: string
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    PatientInput:
      type: object
      required: [fullName, age, gender, mobileNumber, email]
      properties:
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string

    PatientUpdate:
      type: object
      properties:
        fullName:
          type: string
        age:
          type: integer
        gender:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string

    PatientListResponse:
      type: object
      required: [patients, total, page, limit]
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    OrthoCase:
      type: object
      required: [id, caseCode, patientId, status, createdAt]
      properties:
        id:
          type: integer
        caseCode:
          type: string
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        title:
          type: string
        status:
          type: string
          enum: [new, scan_uploaded, analysis_completed, treatment_planning, approved, manufacturing]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    CaseInput:
      type: object
      required: [patientId, title]
      properties:
        patientId:
          type: integer
        title:
          type: string
        notes:
          type: string

    CaseUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [new, scan_uploaded, analysis_completed, treatment_planning, approved, manufacturing]
        notes:
          type: string

    CaseListResponse:
      type: object
      required: [cases, total, page, limit]
      properties:
        cases:
          type: array
          items:
            $ref: "#/components/schemas/OrthoCase"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    Scan:
      type: object
      required: [id, caseId, fileName, fileType, fileSize, jawType, createdAt]
      properties:
        id:
          type: integer
        caseId:
          type: integer
        fileName:
          type: string
        originalName:
          type: string
        fileType:
          type: string
          enum: [stl, obj, ply]
        fileSize:
          type: integer
        jawType:
          type: string
          enum: [upper, lower, both]
        filePath:
          type: string
        createdAt:
          type: string

    DashboardStats:
      type: object
      required: [totalPatients, activeCases, totalCases, newCasesThisMonth, scansUploaded]
      properties:
        totalPatients:
          type: integer
        activeCases:
          type: integer
        totalCases:
          type: integer
        newCasesThisMonth:
          type: integer
        scansUploaded:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        patientName:
          type: ["string", "null"]
        caseCode:
          type: ["string", "null"]
        createdAt:
          type: string

    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer