# lib/api-zod/src/generated/types/caseListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrthoVision Treatment Planning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrthoCase } from './orthoCase';

export interface CaseListResponse {
  cases: OrthoCase[];
  total: number;
  page: number;
  limit: number;
}