# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrthoVision Treatment Planning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt?: string;
}