# lib/api-zod/src/generated/types/orthoCase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrthoVision Treatment Planning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrthoCaseStatus } from './orthoCaseStatus';

export interface OrthoCase {
  id: number;
  caseCode: string;
  patientId: number;
  /** @nullable */
  patientName?: string | null;
  title?: string;
  status: OrthoCaseStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}