# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrthoVision Treatment Planning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  patientCode: string;
  fullName: string;
  age: number;
  gender: string;
  mobileNumber: string;
  email: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}