# lib/api-zod/src/generated/types/statusCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrthoVision Treatment Planning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusCount {
  status: string;
  count: number;
}